v0 Full Grid Implementation (#379)
Training isn't happening properly yet but the full learning pipeline is
there.

Definitely some edge cases, but demo-able.

Unity can be started in two modes, learning or not.  Not learning works
as it does now.  When you call grid.learn or grid.results we query the
bygone server to upload a job / get the results of a job.

Learn mode will poll bygone for jobs, do them, and upload the results.